requirements.txt
flask
numpy
pandas
scipy
scikit-learn
matplotlib
torch
tqdm
psycopg2-binary
pyyaml
wfdb
neurokit2>=0.2.0
confluent-kafka>=2.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0